ip: add ip_darwin / ip_linux files
This ensures that the IP package is not compiled with the netlink dependency in
Darwin (Mac) environments. Packages are imported into the Ginkgo tests which
transitively imported `pkg/ip`, which caused running the Ginkgo tests locally to
not work:

```
$ K8S_VERSION=1.15 ginkgo --focus=K8sServicesTest -v -- -cilium.holdEnvironment=true -cilium.skipLogs=true -cilium.provision=false
Failed to compile test:

../pkg/ip/ip.go:781:36: undefined: netlink.FAMILY_ALL

Ginkgo ran 1 suite in 6.686508356s
Test Suite Failed
```

Factoring out the functions which depend on the netlink library into
a separate file for Linux / Darwin works around this problem.

Signed-off by: Ian Vernon <ian@cilium.io>
