+++ pkg/ip/ip.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Authors of Cilium
+// Copyright 2017-2019 Authors of Cilium
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,9 +21,6 @@ import (
 	"math/big"
 	"net"
 	"sort"
-	"strings"
-
-	"github.com/vishvananda/netlink"
 )
 
 const (
@@ -754,39 +751,6 @@ func initPrivatePrefixes() {
 }
 
 var excludedIPs []net.IP
-
-func initExcludedIPs() {
-	// We exclude below bad device prefixes from address selection ...
-	prefixes := []string{
-		"docker",
-	}
-	links, err := netlink.LinkList()
-	if err != nil {
-		return
-	}
-	for _, l := range links {
-		// ... also all down devices since they won't be reachable.
-		if l.Attrs().OperState == netlink.OperUp {
-			skip := true
-			for _, p := range prefixes {
-				if strings.HasPrefix(l.Attrs().Name, p) {
-					skip = false
-					break
-				}
-			}
-			if skip {
-				continue
-			}
-		}
-		addr, err := netlink.AddrList(l, netlink.FAMILY_ALL)
-		if err != nil {
-			continue
-		}
-		for _, a := range addr {
-			excludedIPs = append(excludedIPs, a.IP)
-		}
-	}
-}
 
 func init() {
 	initPrivatePrefixes()
